Add closing logic to SasMux, refactor AmqpConnection
